also persist the column widths in list-m2m
